% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_.R
\name{check_}
\alias{check_}
\alias{check_class}
\alias{check_value}
\title{Functions to check things}
\usage{
check_class(object, needed)

check_value(object, needed)
}
\arguments{
\item{object}{object to test}

\item{needed}{vector of class that are searched for in an object}
}
\value{
If passes, function do not return anything; if not they return an error
}
\description{
Functions to check things
}
\details{
\code{check_class} checking for specific class. In that case \code{needed} is a vector of class that are searched for in an object
\code{check_values} checking for specific values. In that case \code{needed} is a named list. Name of list indicates method of comparison (values, regex, range)
}
\examples{
check_class("abba", "character")
check_class(c(1:3), "integer")
check_value(c(1:5), list(range = c(-6, 6)))
}
