% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enclave.R
\name{enclave_}
\alias{enclave_}
\alias{enclave_lst}
\alias{enclave_dtl}
\alias{enclave_vec}
\title{Set of functions for adjusting some of polygons for being surrounded completely ('enclave') by other polygon}
\usage{
enclave_lst()

enclave_dtl(lstOb = enclave_lst())

enclave_vec(vecPolyId)
}
\arguments{
\item{lstOb}{list returned by enclave_lst()}

\item{vecPolyId}{vector of polygons id}
}
\value{
each function serves different purpose and returnes different object:
\itemize{
 \item{enclave_lst()}{returns named list. Each element is a vector of ids of polygons: one is an enclave, other is polygon surrounding it. Name of element is an id of that set}
 \item{enclave_dtl()}{returns data.table with columns 'JPT_KOD_JE' and 'id'. 'id' contains an id of set of polygons, and 'JPT_KOD_JE' its actual ids}
 \item{enclave_vec()}{returns a vector that switches original ids of polygons. If an id is listed in enclave_dtl()$JPT_KOD_JE, it gets corresponding enclave_dtl()$id}
}
}
\description{
Set of functions for adjusting some of polygons for being surrounded completely ('enclave') by other polygon
}
\examples{
enclave_lst()
enclave_dtl()
enclave_vec(vecPolyId = c("2065", "1411"))
}
