% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_polygons.R
\name{merge_polygons}
\alias{merge_polygons}
\title{Merging polygons}
\usage{
merge_polygons(
  orPol,
  uklad,
  clusterName = "klaster",
  polysId = "id",
  geometryName = "g"
)
}
\arguments{
\item{orPol}{oryginalny shape}

\item{uklad}{data.table o kolumnie 'id' (ktora ma odpowiadac 'JPT_KOD_JE') i klaster czyli id klastra}

\item{clusterName}{column name in uklad that corresponds to cluster id}

\item{polysId}{column name in uklad that corresponds to polygon id}

\item{geometryName}{column name with geometry}
}
\value{
sf object?
}
\description{
Merging polygons
}
\details{
zamysl jest taki, ze dla jednego klastra jest jedno id klastra. dla jednego id klastra jest wiele id powiatow. dla jednego id klastra
}
\examples{
merge_polygons(shapes_wojewodztwa, data.table(id = c('02', '16'), klaster = "1"))
}
