% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neigh_list}
\alias{neigh_list}
\title{A sf object of simplified counties}
\format{
A list of data.tables. Each data.table
\describe{
  \item{ID_TER}{NUTS code of territory}
  \item{ID_NEIGH}{NUTS of its neighbour}
}
}
\source{
\url{https://gis-support.pl/baza-wiedzy-2/dane-do-pobrania/granice-administracyjne/}
}
\usage{
neigh_list
}
\description{
A sf object of simplified counties
}
\keyword{datasets}
