% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwarz.R
\name{obwarz}
\alias{obwarz}
\alias{obwarz_list}
\alias{obwarz_dt}
\title{Set of functions for adjusting some of polygons for being surrounded completely by one other polygon}
\usage{
obwarz_list()

obwarz_dt(o = obwarz_list())

obwarz(x)
}
\arguments{
\item{o}{list returned by obwarz_list()}

\item{x}{vector of polygons id}
}
\value{
obwarz_list returns named list
\item{obwarz_list()} returns named list
\item{obwarz_dt()} returns data.table
\item{obwarz()} returns a vector that switches original ids for that surounding them
}
\description{
Set of functions for adjusting some of polygons for being surrounded completely by one other polygon
}
\examples{
obwarz_list()
obwarz_dt()
obwarz(x = c("2065", "1411"))
}
