% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shapes_gminy}
\alias{shapes_gminy}
\title{A sf object of simplified counties}
\format{
A sf data.frame object with columns such as
\describe{
  \item{JPT_KOD_JE}{NUTS2 code '260204', '2602063'}
  \item{JPT_NAZWA_}{name of NUTS2}
  \item{Shape_Area}{area of shape}
  \item{geometry}{geometry of shape}
}
}
\source{
\url{https://gis-support.pl/baza-wiedzy-2/dane-do-pobrania/granice-administracyjne/}
}
\usage{
shapes_gminy
}
\description{
A sf object of simplified counties
}
\keyword{datasets}
