% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shapes_wojewodztwa}
\alias{shapes_wojewodztwa}
\title{A sf object of simplified voidships}
\format{
A sf data.frame object with columns such as
\describe{
  \item{JPT_KOD_JE}{NUTS2 code '02', '32'}
  \item{JPT_NAZWA_}{name of NUTS2}
  \item{Shape_Area}{area of shape}
  \item{geometry}{geometry of shape}
}
}
\source{
\url{https://gis-support.pl/baza-wiedzy-2/dane-do-pobrania/granice-administracyjne/}
}
\usage{
shapes_wojewodztwa
}
\description{
A sf object of simplified voidships
}
\keyword{datasets}
