% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djikstra.R
\name{short_paths}
\alias{short_paths}
\title{Calculates shortest paths from start to every othe point in given graph}
\usage{
short_paths(start, g)
}
\arguments{
\item{start}{id}

\item{g}{data.table with ID_TER, ID_NEIGH and odl}
}
\value{
data.table with columns: 'a' id of polys, 'odl' (dist from start) and 'b' (path from start to end, excludin starting point)
}
\description{
g might be a product of table_dists
}
\examples{
short_paths(start = '02', g = table_dists('wojewodztwa'))
}
