% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp_centroids.R
\name{shp_centroids}
\alias{shp_centroids}
\title{Returns data.table with centroids of polygon}
\usage{
shp_centroids(shp)
}
\arguments{
\item{shp}{shapefile object}
}
\value{
data.table with columns 'id', 'x' and 'y'
}
\description{
Returns data.table with centroids of polygon
}
\examples{
shp_centroids(sfpu::shapes_wojewodztwa)
}
