% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djikstra.R
\name{table_dists}
\alias{table_dists}
\title{Creates data.table of distances between neighbouring centroids}
\usage{
table_dists(lvl, neighs = sfpu::neigh_list[[lvl]])
}
\arguments{
\item{lvl}{lvl (wojewodztwa, powiaty, gminy)}

\item{neighs}{data.table of neighbours}
}
\value{

}
\description{
Creates data.table of distances between neighbouring centroids
}
\examples{
table_dists("wojewodztwa")
}
