% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djikstra.R
\name{vis_paths}
\alias{vis_paths}
\title{Visualizes shortest path}
\usage{
vis_paths(path, end)
}
\arguments{
\item{path}{reuslt of short_paths()}

\item{end}{id of ending polygon}
}
\value{
ggplot2 object with visualization of path
}
\description{
Visualizes shortest path
}
\examples{
oj <- short_paths(start = '12', g = table_dists('wojewodztwa'))
vis_paths(path = oj, end = "32")
}
