% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_out_neigbours.R
\name{find_out_neigbours}
\alias{find_out_neigbours}
\title{Finding out neighbouring polygons}
\usage{
find_out_neigbours(poly, kol, idName = "co", neighName = "s")
}
\arguments{
\item{poly}{sf object with polygons}

\item{kol}{column name}

\item{idName}{name of column with id}

\item{neighName}{name of column with neighbours id}
}
\value{

}
\description{
Finding out neighbouring polygons
}
\examples{
find_out_neigbours(poly = shapes_wojewodztwa, kol = "JPT_KOD_JE")

}
