% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_polygons.R
\name{merge_polygons}
\alias{merge_polygons}
\title{Merging polygons}
\usage{
merge_polygons(
  orPol,
  uklad,
  clusterName = "klaster",
  polysId = "id",
  geometryName = "g"
)
}
\arguments{
\item{orPol}{shapefile with needed geometry}

\item{uklad}{data.table with a column with polygons 'id' (which has to correspond to 'JPT_KOD_JE' in shapefiles) and cluster id}

\item{clusterName}{column name in \code{uklad} that corresponds to cluster id}

\item{polysId}{column name in \code{uklad} that corresponds to polygon id}

\item{geometryName}{column name with geometry}
}
\value{
sf object with columns containing cluster name (provided via \code{clusterName}) and new geometry of a cluster (with a name provided via \code{geometryName})
}
\description{
Merging polygons
}
\details{
for each cluster there is a one cluster id. for each cluster id there are >=1 id of a polygon
to do: if a cluster has only one polygon, there is no need to merge
}
\examples{
res <- merge_polygons(shapes_wojewodztwa, data.table(id = c('02', '16'), klaster = "1"))
ggplot2::ggplot(res) + ggplot2::geom_sf()
}
