% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djikstra.R
\name{short_paths}
\alias{short_paths}
\title{Calculating shortest path is based on Djikstra algorithm}
\usage{
short_paths(start, g)
}
\arguments{
\item{start}{id of a polygon that is a start}

\item{g}{data.table with columns 'ID_TER', 'ID_NEIGH' and 'odl'}
}
\value{
data.table with columns: 'a' id of polygons, 'odl' (dist from start in meters) and 'b' (ids of polygon in exact order that constructs shortest path from \code{start} to \code{a}, excluding starting point, in the form of id1 - id2 - ... - idN)
}
\description{
Calculates shortest paths from start to every other point in given graph using Djikstra algorithm
}
\details{
it is recommended that \code{g} be a product of \code{table_dists()}
}
\examples{
short_paths(start = '02', g = table_dists('wojewodztwa'))
}
