% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djikstra.R
\name{table_dists}
\alias{table_dists}
\title{Calculating distances between adjacent centroids}
\usage{
table_dists(lvl, neighs = sfpu::neigh_list[[lvl]])
}
\arguments{
\item{lvl}{lvl (one of 'wojewodztwa', 'powiaty' or 'gminy')}

\item{neighs}{data.table containing pairs ids of polygons (default: adjacent polygons of level \code{lvl})}
}
\value{
data.table with columns 'ID_TER', 'ID_NEIGH' and 'odl' (euclidean distance between centroids of polygons in 'ID_TER' and 'ID_NEIGH')
}
\description{
Creates data.table of distances between centroids of polygons
}
\details{
\code{neighs} must have columns 'ID_TER' and 'ID_NEIGH'
}
\examples{
table_dists("wojewodztwa")
}
