% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djikstra.R
\name{vis_paths}
\alias{vis_paths}
\title{Visualizes shortest path}
\usage{
vis_paths(path, end)
}
\arguments{
\item{path}{a result of \code{sfpu::short_paths}}

\item{end}{an id of ending polygon}
}
\value{
ggplot2 object with visualization of a path
}
\description{
Visualizes shortest path based on \code{path}s previously calculated
}
\examples{
distances_between_voids <- table_dists('wojewodztwa')
shortest_paths_from_12_void <- short_paths(start = '12', g = distances_between_voids)
vis_paths(path = shortest_paths_from_12_void, end = "32")
}
